refactor(api): make generating of blog posts using Gemini reusable
